src/requirements.txt
fastapi[all]==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.25
asyncpg==0.29.0
passlib[bcrypt,argon2]==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings==2.1.0
jinja2==3.1.4
python-multipart==0.0.9
httpx==0.27.0
